+++ lib/rubocop/cop/style/struct_inheritance.rb
@@ -8,12 +8,20 @@ module RuboCop
       # @example
       #   # bad
       #   class Person < Struct.new(:first_name, :last_name)
+      #     def age
+      #       42
+      #     end
       #   end
       #
       #   # good
-      #   Person = Struct.new(:first_name, :last_name)
+      #   Person = Struct.new(:first_name, :last_name) do
+      #     def age
+      #       42
+      #     end
+      #   end
       class StructInheritance < Cop
-        MSG = "Don't extend an instance initialized by `Struct.new`.".freeze
+        MSG = "Don't extend an instance initialized by `Struct.new`. " \
+              'Use a block to customize the struct.'.freeze
 
         def on_class(node)
           _name, superclass, _body = *node