+++ spec/rubocop/cop/style/struct_inheritance_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe RuboCop::Cop::Style::StructInheritance do
   it 'registers an offense when extending instance of Struct' do
     expect_offense(<<-RUBY.strip_indent)
       class Person < Struct.new(:first_name, :last_name)
-                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Don't extend an instance initialized by `Struct.new`.
+                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Don't extend an instance initialized by `Struct.new`. Use a block to customize the struct.
       end
     RUBY
   end
@@ -14,7 +14,7 @@ RSpec.describe RuboCop::Cop::Style::StructInheritance do
   it 'registers an offense when extending instance of Struct with do ... end' do
     expect_offense(<<-RUBY.strip_indent)
       class Person < Struct.new(:first_name, :last_name) do end
-                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Don't extend an instance initialized by `Struct.new`.
+                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Don't extend an instance initialized by `Struct.new`. Use a block to customize the struct.
       end
     RUBY
   end
@@ -35,5 +35,15 @@ RSpec.describe RuboCop::Cop::Style::StructInheritance do
 
   it 'accepts assignment to Struct.new' do
     expect_no_offenses('Person = Struct.new(:first_name, :last_name)')
+  end
+
+  it 'accepts assignment to block form of Struct.new' do
+    expect_no_offenses(<<-RUBY.strip_indent)
+      Person = Struct.new(:first_name, :last_name) do
+        def age
+          42
+        end
+      end
+    RUBY
   end
 end